Create HttpBasicPreAuthenticatedListener and factory class
This listener grants compatibility with mod_auth_ldap for Apache, as the HTTP password server variables will contain the username/password of the successfully-authenticated LDAP user.  Down the line, it may make sense to also implement a listener more like the HTTP basic listener (not pre-auth), which will attempt a bind against LDAP to verify the credentials.

Note that likely due to a shortcoming of Symfony2's security component, custom authentication listeners cannot be registered by the security extension.  A temporary solution is to abuse the stock X509 listener to support HTTP pre-authentication.  Specifying the ldap user provider and employing x509 in your firewall, the following can be added in your configuration file to force X509 to read HTTP auth server variables in place of the usual SSL credentials:

    parameters:
        security.authentication.x509.user: PHP_AUTH_USER
        security.authentication.x509.credentials: PHP_AUTH_PW